pyproject.toml
[tool.poetry]
name = "steamtrackerbot"
version = "0.0.0"
description = "Steam Tracker Discord bot"
authors = ["Hhummann <hhummannv@gmail.com>"]
license = "Jade Wave"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-steam-api = "^1.2.2"
discord-py = "^2.3.2"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
